Remove MUI components from RecipeCard
