+++ src/components/App/App.scss
@@ -1,5 +1,11 @@
+
 * {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
+}
+
+.App {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
 }
