+++ src/main.scss
@@ -1,2 +1,8 @@
 @import './containers/IngredientsForm/IngredientsForm.scss';
 @import './components/App/App.scss';
+body {
+  background-image: url("./assets/board.jpg");
+  background-size: 100% auto, cover;
+  // background-position: center;
+  background-repeat: no-repeat;
+}
